Fix private variable name and string value
This fixes the variable `self._password` which is supposed to be
`self.__password` as defined in the setter and getter methods. Also it
fixes the error in the the string casing when checking for password
validity in the `is_valid_password()` method. Initially, the the hashed
value is converted to uppercase, then checked with the value in
`self.__password` variable. This wasn't working because the original
value saved in `self.__password` is in lowercase, changing it solved the
problem.